python/shared/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-clip-saver-shared"
version = "1.0.0"
description = "Shared database and utilities for Discord Clip Saver"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tortoise-orm>=0.20.0,<1.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "aerich>=0.7.2,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.aerich]
tortoise_orm = "shared.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
